+++ CHANGELOG.md
@@ -8,8 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Implemented Steering Service.
 
 ### Changed
+- Peloton data only logs every second. 
 
 ### Hardware